requirements.txt
langchain>=0.0.267
langgraph>=0.0.17
groq>=0.3.0
faiss-cpu>=1.7.4
chromadb>=0.4.15
python-dotenv>=1.0.0
textual>=0.38.1
playwright>=1.40.0
tiktoken>=0.5.1
pydantic>=2.4.2
redis>=5.0.0
numpy>=1.24.3
tqdm>=4.65.0 